Add solution for revenge of the pancakes
This is my solution for revenge of the pancakes. Input is received from
the user in the input.go file inside the cmd/rop package. The kitchen
will turn the input string ("+-+") into a stack of pancakes. When given
to the waiter, he will sort them and return the number of flips needed
to make every pancake face up.